requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
aiohttp==3.9.1
requests==2.31.0
numpy==1.24.3
pandas==2.1.3
pillow==10.1.0
opencv-python==4.8.1.78
groq==0.4.1
huggingface-hub==0.19.4
gitpython==3.1.40
PyGithub==2.1.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
httpx==0.25.2
